Move email link to external file
