+++ packages/template-retail-react-app/app/ssr.js
@@ -22,84 +22,9 @@ import {defaultPwaKitSecurityHeaders} from '@salesforce/pwa-kit-runtime/utils/mi
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 import helmet from 'helmet'
 
-import crypto from 'crypto'
 import express from 'express'
-import https from 'https'
 
-/*
- Make a POST request using native https module, wrapped in a Promise with JSON
- encode and decode
-*/
-function asyncJsonHttpsPost(options, postObject) {
-    return new Promise((resolve, reject) => {
-        const req = https.request(options, (response) => {
-            let data = '';
-
-            response.on('data', (chunk) => {
-                data += chunk;
-            });
-
-            response.on('end', () => {
-                resolve(JSON.parse(data));
-            });
-        });
-
-        req.on('error', (error) => {
-            reject(error);
-        });
-
-        req.write(JSON.stringify(postObject));
-
-        req.end();
-    });
-}
-
-async function emailLink(emailId, templateId, magicLink) {
-
-      function generateUniqueId() {
-        return crypto.randomBytes(16).toString('hex');
-      }
-
-      const tokenOptions = {
-        method: 'POST',
-        host: `${process.env.MARKETING_CLOUD_SUBDOMAIN}.auth.marketingcloudapis.com`,
-        path: '/v2/token',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      };
-
-      const tokenBody = {
-        grant_type: 'client_credentials',
-        client_id: process.env.MARKETING_CLOUD_CLIENT_ID,
-        client_secret: process.env.MARKETING_CLOUD_CLIENT_SECRET
-      }
-
-      const token = (await asyncJsonHttpsPost(tokenOptions, tokenBody)).access_token
-
-      const emailOptions = {
-        method: 'POST',
-        host: `${process.env.MARKETING_CLOUD_SUBDOMAIN}.rest.marketingcloudapis.com`,
-        path: `/messaging/v1/email/messages/${generateUniqueId()}`,
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`,
-        },
-      };
-
-      const emailBody = {
-        definitionKey: templateId,
-        recipient: {
-          contactKey: emailId,
-          to: emailId,
-          attributes: { 'magic-link': magicLink },
-        },
-      }
-        
-      const emailResponse = await asyncJsonHttpsPost(emailOptions, emailBody)
-
-      return emailResponse
-}
+import { emailLink } from './marketing-cloud-email-link'
 
 const options = {
     // The build directory (an absolute path)
@@ -190,9 +115,8 @@ const {handler} = runtime.createHandler(options, (app) => {
             token
         } = req.body
         const magicLink = `${base}/reset-password?email=${email_id}&token=${token}`
-        res.send(magicLink)
-        //const emailLinkResponse = await emailLink(email_id, process.env.MARKETING_CLOUD_RESET_PASSWORD_TEMPLATE, magicLink)
-        //res.send(emailLinkResponse)
+        const emailLinkResponse = await emailLink(email_id, process.env.MARKETING_CLOUD_RESET_PASSWORD_TEMPLATE, magicLink)
+        res.send(emailLinkResponse)
     })
 
     app.get('/robots.txt', runtime.serveStaticFile('static/robots.txt'))
